rename task-plugin to datatransfer-plugin
