+++ server/odc-service/pom.xml
@@ -110,7 +110,7 @@
         </dependency>
         <dependency>
             <groupId>com.oceanbase</groupId>
-            <artifactId>task-plugin-api</artifactId>
+            <artifactId>datatransfer-plugin-api</artifactId>
         </dependency>
         <dependency>
             <groupId>org.pf4j</groupId>
@@ -259,7 +259,7 @@
         </dependency>
         <dependency>
             <groupId>com.oceanbase</groupId>
-            <artifactId>task-plugin-ob-mysql</artifactId>
+            <artifactId>datatransfer-plugin-ob-mysql</artifactId>
             <version>4.2.1-SNAPSHOT</version>
             <scope>compile</scope>
         </dependency>
