+++ server/plugins/datatransfer-plugin-mysql/src/main/assembly/assembly.xml
