+++ server/plugins/datatransfer-plugin-ob-mysql/src/test/java/com/oceanbase/odc/plugin/task/obmysql/datatransfer/LoadParameterFactoryTest.java
