+++ server/plugins/datatransfer-plugin-ob-mysql/src/test/resources/datatransfer/data/CHECKPOINT.bin
