+++ server/plugins/datatransfer-plugin-ob-mysql/src/test/resources/datatransfer/data/GSH/TABLE/TEST-schema.sql
