+++ server/plugins/datatransfer-plugin-ob-mysql/src/test/resources/datatransfer/data/MANIFEST.bin
