+++ server/plugins/datatransfer-plugin-ob-oracle/src/main/assembly/assembly.xml
