+++ server/plugins/pom.xml
@@ -22,10 +22,10 @@
         <module>schema-plugin-ob-oracle</module>
         <module>schema-plugin-mysql</module>
         <module>schema-plugin-odp-sharding-ob-mysql</module>
-        <module>task-plugin-api</module>
-        <module>task-plugin-ob-mysql</module>
-        <module>task-plugin-ob-oracle</module>
-        <module>task-plugin-mysql</module>
+        <module>datatransfer-plugin-api</module>
+        <module>datatransfer-plugin-ob-mysql</module>
+        <module>datatransfer-plugin-ob-oracle</module>
+        <module>datatransfer-plugin-mysql</module>
     </modules>
     <properties>
         <maven-assembly-plugin.version>2.4</maven-assembly-plugin.version>
@@ -71,7 +71,7 @@
             </dependency>
             <dependency>
                 <groupId>com.oceanbase</groupId>
-                <artifactId>task-plugin-ob-mysql</artifactId>
+                <artifactId>datatransfer-plugin-ob-mysql</artifactId>
                 <version>${project.version}</version>
                 <scope>provided</scope>
             </dependency>
