+++ server/starters/desktop-starter/pom.xml
@@ -33,7 +33,7 @@
         </dependency>
         <dependency>
             <groupId>com.oceanbase</groupId>
-            <artifactId>task-plugin-api</artifactId>
+            <artifactId>datatransfer-plugin-api</artifactId>
         </dependency>
     </dependencies>
     <build>
